Fix compilation errors when compiling tests in release. (#245)

+++ platform/darwin/test/MGLTestAssertionHandler.m
@@ -35,9 +35,9 @@
                                        );
     }
     else {
-        NSLog(@"Assertion Failure: %@:%lu: %@ - %@",
+        NSLog(@"Assertion Failure: %@:%ld: %@ - %@",
               fileName,
-              line,
+              (long)line,
               condition,
               description);
     }
@@ -66,9 +66,9 @@
                                        description);
     }
     else {
-        NSLog(@"Assertion Failure: %@:%lu: %@ - %@",
+        NSLog(@"Assertion Failure: %@:%ld: %@ - %@",
               fileName,
-              line,
+              (long)line,
               condition,
               description);
     }
